Update the derived stream generators
- normalize the main/crash summary 'os'
- rename the existing executive_summary to *_full since in handles the
  the s3 full telemetry stream
- add an executive_summary generator to handle the s3 executive summary
  stream
